[Update] fix inference bugs
