+++ nkyolo/engine/predictor.py
@@ -49,7 +49,7 @@ from nkyolo.utils.files import increment_path
 
 STREAM_WARNING = """
 WARNING ⚠️ inference results will accumulate in RAM unless `stream=True` is passed, causing potential out-of-memory
-errors for large sources or long-running streams and videos. See https://docs.ultralytics.com/modes/predict/ for help.
+errors for large sources or long-running streams and videos. 
 
 Example:
     results = model(source=..., stream=True)  # generator of Results objects
